move lib folder
