+++ patches/activate-cross-signing-and-secure-storage-react/matrix-react-sdk+3.71.1.patch
@@ -1,5 +1,5 @@
 diff --git a/node_modules/matrix-react-sdk/src/components/structures/MatrixChat.tsx b/node_modules/matrix-react-sdk/src/components/structures/MatrixChat.tsx
-index 575ff07..782cdd1 100644
+index 575ff07..3424e04 100644
 --- a/node_modules/matrix-react-sdk/src/components/structures/MatrixChat.tsx
 +++ b/node_modules/matrix-react-sdk/src/components/structures/MatrixChat.tsx
 @@ -32,6 +32,8 @@ import { throttle } from "lodash";
@@ -7,7 +7,7 @@ index 575ff07..782cdd1 100644
  import { RoomType } from "matrix-js-sdk/src/@types/event";
  import { DecryptionError } from "matrix-js-sdk/src/crypto/algorithms";
 +//:tchap: import incoming key request handler from tchap project
-+import KeyRequestHandler from '../../../../../src/lib/IncomingKeyRequestHandler';
++import KeyRequestHandler from '../../../../../src/tchap/lib/IncomingKeyRequestHandler';
  
  // focus-visible is a Polyfill for the :focus-visible CSS pseudo-attribute used by various components
  import "focus-visible";
