+++ src/components/views/rooms/TchapExternalRoomHeader.tsx
@@ -3,7 +3,7 @@ import { useTypedEventEmitter } from "matrix-react-sdk/src/hooks/useEventEmitter
 import React, { useState } from "react";
 import { _t } from "matrix-react-sdk/src/languageHandler";
 
-import { TchapRoomAccessRule, TchapRoomAccessRulesEventId } from "../../../@types/tchap";
+import { TchapRoomAccessRule, TchapRoomAccessRulesEventId } from "../../../tchap/@types/tchap";
 import TchapRoomUtils from "../../../util/TchapRoomUtils";
 
 export interface IProps {
