+++ src/tchap/lib/createTchapRoom.ts
@@ -2,7 +2,7 @@ import { IOpts } from "matrix-react-sdk/src/createRoom";
 import { ICreateRoomOpts } from "matrix-js-sdk/src/@types/requests";
 import { HistoryVisibility, JoinRule, Preset, Visibility } from "matrix-js-sdk/src/@types/partials";
 
-import { TchapRoomAccessRule, TchapRoomAccessRulesEventId, TchapRoomType } from "../../@types/tchap";
+import { TchapRoomAccessRule, TchapRoomAccessRulesEventId, TchapRoomType } from "../@types/tchap";
 
 export const DEFAULT_FEDERATE_VALUE = true;
